# examples/nest/your-app/swagger/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/register": {
      "post": {
        "description": "Creates a user. Email must be unique. Does not log in: call `POST /auth/login` to get a token.",
        "operationId": "auth.register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created user. `avatarUrl` is `null`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid email or password shorter than 6 characters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "A user with this email already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "summary": "Register a new user",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "description": "Exchanges email and password for an access token and returns the user. Every login issues a new token; old tokens stay valid.",
        "operationId": "auth.login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenDto"
                }
              }
            }
          },
          "401": {
            "description": "Wrong email or password.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "summary": "Log in",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "description": "Returns the user the bearer token belongs to.",
        "operationId": "auth.me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Current user",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/whoami": {
      "get": {
        "deprecated": true,
        "description": "Same as `GET /auth/me`. Kept for old clients, use `GET /auth/me` instead.",
        "operationId": "auth.whoami",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Current user (deprecated)",
        "tags": [
          "auth"
        ]
      }
    },
    "/posts": {
      "get": {
        "description": "Returns published posts in creation order. Drafts (`publishedAt: null`) are never listed, even to their author.",
        "operationId": "posts.list",
        "parameters": [
          {
            "name": "authorId",
            "required": false,
            "in": "query",
            "description": "Only return posts of this author.",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PostDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List published posts",
        "tags": [
          "posts"
        ]
      },
      "post": {
        "description": "Creates a draft (`publishedAt: null`) owned by the current user. Publish it with `POST /posts/{id}/publish`.",
        "operationId": "posts.create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDto"
                }
              }
            }
          },
          "400": {
            "description": "Empty title or content.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a post",
        "tags": [
          "posts"
        ]
      }
    },
    "/posts/{id}": {
      "get": {
        "description": "Returns any post, drafts included. No authorization needed.",
        "operationId": "posts.get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Post id.",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDto"
                }
              }
            }
          },
          "404": {
            "description": "No post with this id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "summary": "Get a post by id",
        "tags": [
          "posts"
        ]
      },
      "delete": {
        "description": "Deletes the post and all its comments. Only the author may delete, otherwise 403.",
        "operationId": "posts.remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Post id.",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted. Empty body."
          },
          "401": {
            "description": "Missing or invalid bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "The current user is not the author.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "No post with this id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a post",
        "tags": [
          "posts"
        ]
      }
    },
    "/posts/{id}/publish": {
      "post": {
        "description": "Sets `publishedAt` to the current time. Only the author may publish, otherwise 403. Publishing an already published post keeps the original `publishedAt`.",
        "operationId": "posts.publish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Post id.",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "The current user is not the author.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "No post with this id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Publish a post",
        "tags": [
          "posts"
        ]
      }
    },
    "/posts/{id}/comments": {
      "post": {
        "description": "Adds a comment to any post, drafts included. Pass `parentId` to reply to another comment of the same post.",
        "operationId": "comments.create",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Post id.",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created comment. `replies` is empty.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentDto"
                }
              }
            }
          },
          "400": {
            "description": "Empty text or non-integer `parentId`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "No post with this id, or `parentId` is not a comment of this post.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Comment on a post",
        "tags": [
          "comments"
        ]
      },
      "get": {
        "description": "Returns top-level comments, oldest first. Replies are nested in `replies` at any depth.",
        "operationId": "comments.list",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Post id.",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommentDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No post with this id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "summary": "Comment tree of a post",
        "tags": [
          "comments"
        ]
      }
    },
    "/users/me/avatar": {
      "post": {
        "description": "Sets the avatar of the current user. Any file is accepted; `avatarUrl` becomes `/avatars/<userId><ext>`, where the extension comes from the uploaded file name. Uploading again replaces the avatar.",
        "operationId": "users.uploadAvatar",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "The image file."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The user with the new `avatarUrl`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "400": {
            "description": "No `file` field in the form.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload an avatar",
        "tags": [
          "users"
        ]
      }
    },
    "/test/reset": {
      "post": {
        "description": "Deletes all users, tokens, posts and comments, and restarts ids from 1. Exists only when the server runs with `TEST_MODE=1`.",
        "operationId": "test.reset",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "ok"
                  ],
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Reset all data",
        "tags": [
          "test"
        ]
      }
    }
  },
  "info": {
    "title": "Nest example API",
    "description": "Blog backend: users, posts with drafts, nested comments, avatars. Data lives in memory.",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Unique email. Used as the login.",
            "example": "alice@example.com"
          },
          "password": {
            "type": "string",
            "description": "Password, at least 6 characters.",
            "minLength": 6,
            "example": "secret123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "User id.",
            "example": 1
          },
          "email": {
            "type": "string",
            "description": "Email the user registered with.",
            "example": "alice@example.com"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true,
            "description": "Avatar URL. `null` until the user uploads an avatar.",
            "example": "/avatars/1.png"
          }
        },
        "required": [
          "id",
          "email",
          "avatarUrl"
        ]
      },
      "ErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "description": "HTTP status code.",
            "example": 404
          },
          "message": {
            "description": "Error message. Validation errors (400) return a list: one message per failed constraint.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          },
          "error": {
            "type": "string",
            "description": "HTTP status text, e.g. \"Not Found\".",
            "example": "Not Found"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email the user registered with.",
            "example": "alice@example.com"
          },
          "password": {
            "type": "string",
            "description": "User password.",
            "example": "secret123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "TokenDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Send it as `Authorization: Bearer <accessToken>`."
          },
          "user": {
            "description": "The logged-in user.",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserDto"
              }
            ]
          }
        },
        "required": [
          "accessToken",
          "user"
        ]
      },
      "PostDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Post id.",
            "example": 1
          },
          "title": {
            "type": "string",
            "description": "Post title."
          },
          "content": {
            "type": "string",
            "description": "Post body."
          },
          "authorId": {
            "type": "number",
            "description": "Id of the user who created the post."
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "When the post was published. `null` while the post is a draft."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the post was created."
          }
        },
        "required": [
          "id",
          "title",
          "content",
          "authorId",
          "publishedAt",
          "createdAt"
        ]
      },
      "CreatePostDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Post title, non-empty.",
            "example": "Hello"
          },
          "content": {
            "type": "string",
            "description": "Post body, non-empty.",
            "example": "First post"
          }
        },
        "required": [
          "title",
          "content"
        ]
      },
      "CreateCommentDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "description": "Comment text, non-empty.",
            "example": "Nice post!"
          },
          "parentId": {
            "type": "number",
            "description": "Id of the comment to reply to. It must belong to the same post. Omit for a top-level comment."
          }
        },
        "required": [
          "text"
        ]
      },
      "CommentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Comment id.",
            "example": 1
          },
          "text": {
            "type": "string",
            "description": "Comment text."
          },
          "authorId": {
            "type": "number",
            "description": "Id of the user who wrote the comment."
          },
          "replies": {
            "description": "Replies to this comment, oldest first.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommentDto"
            }
          }
        },
        "required": [
          "id",
          "text",
          "authorId",
          "replies"
        ]
      }
    }
  }
}